# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: InvestKE Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /investments:
    get:
      operationId: getMyInvestments
      tags: [investments]
      summary: Get current user's investments
      responses:
        "200":
          description: Investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Start an investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "201":
          description: Investment started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
  /investments/internship:
    post:
      operationId: activateInternship
      tags: [investments]
      summary: Activate 2-day internship package for new users
      responses:
        "201":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "400":
          description: Not eligible or already activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /investments/{id}:
    get:
      operationId: getInvestment
      tags: [investments]
      summary: Get investment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
...
  /otp/send:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP via WhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpSendInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify an OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /otp/send-withdrawal:
    post:
      operationId: sendWithdrawalOtp
      tags: [auth]
      summary: Send OTP for withdrawal (uses user phone on file)
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Plan:
      type: object
      required: [id, name, minDeposit, dailyReturnPercent, durationDays, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        minDeposit:
          type: number
          description: Fixed plan cost in KES
        dailyReturnPercent:
          type: number
          description: Daily return as percentage
        durationDays:
          type: integer
        isActive:
          type: boolean
        isInternship:
          type: boolean
        internshipFixedEarning:
          type: ["number", "null"]
          description: Fixed KES earning for internship package (200 KES)
        createdAt:
          type: string
          format: date-time
...
    Investment:
      type: object
      required: [id, userId, planId, amountInvested, totalEarned, status, startedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        plan:
          $ref: "#/components/schemas/Plan"
        amountInvested:
          type: number
        dailyEarning:
          type: number
        totalEarned:
          type: number
        expectedTotal:
          type: number
        progressPercent:
          type: number
        status:
          type: string
          enum: [active, completed, cancelled]
        startedAt:
          type: string
          format: date-time
        completesAt:
          type: ["string", "null"]
          format: date-time
        lastEarningAt:
          type: ["string", "null"]
          format: date-time
    InvestmentInput:
      type: object
      required: [planId, amount]
      properties:
        planId:
          type: integer
        amount:
          type: number
...
    OtpSendInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
        reason:
          type: string
        ip:
          type: string
        location:
          type: string
    OtpVerifyInput:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
        code:
          type: string
    OtpVerifyResponse:
      type: object
      required: [ok, verified]
      properties:
        ok:
          type: boolean
        verified:
          type: boolean
...